refactor: move export logic to Prompts context
Move evolution metrics serialization logic from controller to
Aludel.Prompts.Evolution.Export module following architecture
principle of keeping business logic outside controllers/views/adapters.

- Created Aludel.Prompts.Evolution.Export with to_json/2 and to_csv/1
- Updated ExportController.evolution/2 to delegate serialization
- Removed duplicate helper functions from controller
